refactor(reservation): 테스트 클래스 이름 및 패키지 구조 정리
- 기능 패키지 기준으로 테스트 코드 위치 정리
- 단위 테스트와 통합 테스트 패키지 구분 반영
- 테스 트대상에 맞게 클래스명 정리
- 기존 테스트 fixture 위치 정리
